+++ changelog.md
@@ -1,4 +1,7 @@
 **Bug Fixes**:
 
 - Fix Adventure Serializer failing completely when serialization fails, causing game glitches
-- 
+
+**New Features**:
+
+- Allow commands to send feedback as chat messages, if command feedback is disabled